% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squarepower.R
\name{compund}
\alias{compund}
\alias{interest}
\title{finding compund interest}
\usage{
interest(p, i, n)
}
\arguments{
\item{p}{is principal amount}

\item{i}{is annual interest rate}

\item{n}{is number of compound periods for a year}
}
\value{
the compund interest
}
\description{
finding compund interest
}
