% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squarepower.R
\name{square}
\alias{square}
\title{testing example square number}
\usage{
square(x)
}
\arguments{
\item{x}{A numeric number to be squared}
}
\value{
the number square from input
}
\description{
input new number to square it.
}
